Collecting messages and preparing them for processing.
The Receiver class implements listening and recording of all messages.
The MainHandler class implements basic processing and preparation for other handlers.
Created module with data types for the bot. UpdateDt(message) type is added to it.
A module with user exceptions is created.
The main.py file launches the bot.