Adding new handlers and Middlewares.
The data field has been added to the UpdateDt data type, it can store arbitrary data.
Reworked ThrottlingMiddleware, now it works correctly.
Added CatchNextMessageMiddleware and handlers for it.
Now you can apply handlers to all messages. To do this, you need to specify the character "*" in the "command" field of the NewCommand class.